requirements.txt
numpy>=1.21.0
matplotlib>=3.4.0
pillow>=8.0.0
scikit-learn>=1.0.0
seaborn>=0.11.0
torch>=1.9.0
torchvision>=0.10.0
scipy>=1.7.0
pandas>=1.3.0
